wip(core): split file code into a module; fix ts-node esm import errors; add types
