+++ nodemon.json
@@ -2,5 +2,5 @@
   "watch": ["src"],
   "ext": ".ts,.js",
   "ignore": [],
-  "exec": "npx ts-node --esm ./src/index.ts"
+  "exec": "npx ts-node --experimental-specifier-resolution=node --esm ./src/index.ts"
 }
