+++ src/types.ts
@@ -1,0 +1,8 @@
+export interface UserRecord {
+  id: string;
+  first_name: string;
+  last_name: string;
+  email: string;
+  department: string;
+  level: string;
+}
