+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
     /* Language and Environment */
     "target": "es2020",
     "lib": ["es2020"],
